backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
mysql-connector-python
python-multipart
PyMySQL
Flask
Flask-Login
Flask-Mail
Flask-Bcrypt
langchain-openai
openai
python-dotenv
google-cloud-secret-manager
google-cloud-storage
Flask-Cors
langchain
qdrant-client
pypdf
python-docx
tiktoken
langchain-community
langgraph
Pillow
